[MIG] purchase_rfq_number: Migration to 17.0
