+++ purchase_rfq_number/__manifest__.py
@@ -4,7 +4,7 @@
 {
     "name": "Purchase For Quotation Numeration",
     "summary": "Different sequence for purchase for quotations",
-    "version": "16.0.1.0.0",
+    "version": "17.0.1.0.0",
     "author": "ProThai, Odoo Community Association (OCA)",
     "license": "AGPL-3",
     "category": "Purchase Management",
