+++ purchase_rfq_number/models/__init__.py
@@ -1,4 +1,5 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
 from . import res_company
+from . import res_config_settings
 from . import purchase_order
