+++ purchase_rfq_number/views/purchase_views.xml
@@ -7,9 +7,8 @@
             <field name="company_id" position="after">
                 <field
                     name="rfq_number"
-                    attrs="{'invisible': ['|', ('rfq_number', '=', 'New'), ('state','not in',('purchase','done'))]}"
+                    invisible="rfq_number == 'New' or state not in ('purchase', 'done')"
                 />
-                <field name="po_number" invisible="1" />
             </field>
         </field>
     </record>
